Don't Move() the results of the XML + SVG builders.
Simplify the XML and SVG builders by removing the Move() Actions.  The faust
compiler already outputs the expected extension for the XML output, so that
Move() basically tried to move the file to itself.  In the case of the SVG
builder this change brings it more in line with what the faust compiler does by
changing the suffix to a lambda that always returns "-svg".  That way SCons
does not prepend the suffix with a dot automatically, and specifying the target
in the FaustSVG() call with a differing suffix does not fall back to the SCons
default suffix.