Do not raise an exception when FAUST is not found.
I believe this was a bad idea.  I think a build system ought to use the exists()
function (or at least check that "FAUST_FAUST" is not empty) and handle the
situation from there.

The assertions are there because of a strange error: when the first Popen()
fails (because faust_faust is None), the second Popen() hangs indefinitely.