SVG2PDF: switch to cairosvg Python module
Switch the SVG2PDF builder from the cairosvg command to its corresponding Python
module.  The reason for this is that the executable checks for the environments
default LOCALE, which is apparently cleared by SCons (or subprocess?) at some
point, leading to the executable failing.

I do hope the API does not change too often...